\name{agilp-package}
\alias{agilp-package}
\alias{agilp}
\docType{package}
\title{
The package contains functions for processing and analysing Agilent express arrays
}
\description{
In this release the function contains a single function, AAProcess. For a description of this function see ?AAProcess.
}
\details{
\tabular{ll}{
Package: \tab agilp\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2010-03-05\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
See above
}
\author{
Benny Chain b.chain@ucl.ac.uk

Maintainer: Benny Chain b.chain@ucl.ac.uk>

}
\references{See documentation under AAProcess function
}

\keyword{ package }
